+++ pyphi/__init__.py
@@ -2,15 +2,15 @@
 # -*- coding: utf-8 -*-
 # __init__.py
 
-#    _______
-#   |__   __|
-#  ____| |____
-# |  __   __  |     ___         ___   __    _
-# | |  | |  | |    / _ \ __ __ / _ \ / /   (_)
-# | |__| |__| |   / ___// // // ___// _ \ / /
-# |____   ____|  /_/    \_, //_/   /_//_//_/
-#    __| |__           /___/
-#   |_______|
+#      _|_|_|
+#        _|
+#  _|_|_|_|_|_|_|    _|_|_|_|    _|      _|  _|_|_|_|    _|      _|  _|_|_|_|_|
+#  _|    _|    _|    _|      _|  _|      _|  _|      _|  _|      _|      _|
+#  _|    _|    _|    _|_|_|_|_|  _|_|_|_|_|  _|_|_|_|_|  _|_|_|_|_|      _|
+#  _|    _|    _|    _|              _|      _|          _|      _|      _|
+#  _|_|_|_|_|_|_|    _|              _|      _|          _|      _|  _|_|_|_|_|
+#        _|
+#      _|_|_|
 
 """
 =====
@@ -57,20 +57,15 @@ See the documentation for the |config| module for a description of the options
 and their defaults.
 """
 
-__title__ = 'pyphi'
-__version__ = '0.7.5'
-__description__ = 'Python library for computing integrated information.',
-__author__ = 'Will Mayner'
-__author_email__ = 'wmayner@gmail.com'
-__author_website__ = 'http://willmayner.com'
-__copyright__ = 'Copyright 2014-2015 Will Mayner'
+from .__about__ import *
+from . import (config, constants, convert, db, examples, jsonify, macro,
+               models, network, node, subsystem, utils, validate)
+from .network import Network
+from .subsystem import Subsystem
 
-
-from . import (network, subsystem, compute, constants, config, db, examples,
-               macro)
-
-Network = network.Network
-Subsystem = subsystem.Subsystem
+__all__ = ['Network', 'Subsystem', 'config', 'constants', 'convert', 'db',
+           'examples', 'jsonify', 'macro', 'models', 'network', 'node',
+           'subsystem', 'utils', 'validate']
 
 import logging
 import logging.config