+++ pyphi/models/concept.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-# concept.py
+# models/concept.py
 
 from collections import namedtuple
 
@@ -395,7 +395,8 @@ class Constellation(tuple):
 
     This is a wrapper around a tuple to provide a nice string representation
     and place to put constellation methods. Previously, constellations were
-    represented as ``tuple(Concept)``; this usage still works in all functions.
+    represented as ``tuple(|Concept|)``; this usage still works in all
+    functions.
     """
 
     def __repr__(self):