+++ test/test_json.py
@@ -2,9 +2,11 @@
 # -*- coding: utf-8 -*-
 # test_json.py
 
+import tempfile
+
 import numpy as np
 
-import pyphi
+from pyphi import compute, jsonify, network
 
 
 def test_jsonify_native():
@@ -20,7 +22,7 @@ def test_jsonify_native():
         'bool': [True, False],
         'null': None
     }
-    assert answer == pyphi.jsonify.loads(pyphi.jsonify.dumps(x))
+    assert answer == jsonify.loads(jsonify.dumps(x))
 
 
 def test_jsonify_numpy():
@@ -36,18 +38,31 @@ def test_jsonify_numpy():
         'np.int64': 2,
         'np.float64': 3.0,
     }
-    assert answer == pyphi.jsonify.loads(pyphi.jsonify.dumps(x))
+    assert answer == jsonify.loads(jsonify.dumps(x))
+
+
+def test_jsonify_network(s):
+    loaded = jsonify.loads(jsonify.dumps(s.network))
+    assert np.array_equal(loaded['tpm'], s.network.tpm)
+    assert np.array_equal(loaded['cm'], s.network.connectivity_matrix)
+    assert loaded['size'] == s.network.size
+
+
+def test_network_from_json(s):
+    f = tempfile.NamedTemporaryFile(mode='wt')
+    jsonify.dump(s.network, f)
+    f.seek(0)
+    assert network.from_json(f.name) == s.network
 
 
 # TODO: these tests need to be fleshed out, they don't do much
 
-
 def test_jsonify_big_mip(s, flushcache, restore_fs_cache):
     flushcache()
-    pyphi.jsonify.loads(pyphi.jsonify.dumps(pyphi.compute.big_mip(s)))
+    jsonify.loads(jsonify.dumps(compute.big_mip(s)))
 
 
 def test_jsonify_complexes(s, flushcache, restore_fs_cache):
     flushcache()
-    complexes = pyphi.compute.complexes(s.network, s.state)
-    pyphi.jsonify.loads(pyphi.jsonify.dumps(complexes))
+    complexes = compute.complexes(s.network, s.state)
+    jsonify.loads(jsonify.dumps(complexes))